requirements.txt
absl-py==0.15.0
alembic==1.4.3
aniso8601==9.0.1
astunparse==1.6.3
Babel==2.9.0
blinker==1.5
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.12.7
chardet==3.0.4
clang==5.0
click==8.1.3
colorama==0.4.6
cycler==0.11.0
Flask==2.1.0
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.1
flatbuffers==1.12
fonttools==4.38.0
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==20.0.4
h5py==3.1.0
idna==2.8
imageio==2.26.0
importlib-metadata==5.2.0
itsdangerous==2.0.0
Jinja2==3.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mysql-connector-python==8.0.26
mysqlclient==2.0.3
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
Pillow==9.3.0
protobuf==3.20.3
psycopg2==2.9.5
psycopg2-binary==2.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
pytz==2022.7
PyWavelets==1.3.0
requests==2.22.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
six==1.15.0
SQLAlchemy==1.3.12
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.0
tensorflow-estimator==2.11.0
termcolor==1.1.0
tifffile==2021.11.2
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==2.0.0
wrapt==1.12.1
zipp==3.11.0
tf2onnx
onnxruntime